requirements.txt
yfinance
matplotlib
openai
streamlit
python-dotenv
pandas
